+++ app/policies/exercise_policy.rb
@@ -1,0 +1,9 @@
+class ExercisePolicy < ApplicationPolicy
+  def index?
+    user.present?
+  end
+
+  def show?
+    user.present?
+  end
+end